DDF-3941 - Extracted the imlpementation of the AuthenticationEndpoint into a service - Created new endpoint under /search/catalog/internal - Changed intrigue code under catalog-search-ui to use new endpoint
