+++ distribution/ddf-common/src/main/resources/security/default.policy
@@ -276,7 +276,7 @@ grant codeBase "file:/admin-core-jolokia/javax.servlet-api/org.apache.shiro.core
     permission java.security.SecurityPermission "insertProvider";
 }
 
-grant codeBase "file:/security-rest-authentication/security-core-api/security-core-impl/security-idp-server/security-idp-client/security-idp-plugin-subjectconfirmation" {
+grant codeBase "file:/security-rest-authn-impl/security-core-api/security-core-impl/security-idp-server/security-idp-client/security-idp-plugin-subjectconfirmation" {
     permission java.io.FilePermission "${ddf.home.perm}etc${/}ws-security${/}server${/}-", "read";
     permission java.io.FilePermission "${ddf.home.perm}etc${/}ws-security${/}issuer${/}-", "read";
     permission java.io.FilePermission "${ddf.home.perm}etc${/}keystores${/}-", "read";
