+++ platform/security/rest/security-rest-authn-service/src/main/java/org/codice/ddf/security/rest/authentication/service/AuthenticationService.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * <p>This is free software: you can redistribute it and/or modify it under the terms of the GNU
+ * Lesser General Public License as published by the Free Software Foundation, either version 3 of
+ * the License, or any later version.
+ *
+ * <p>This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public
+ * License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ */
+package org.codice.ddf.security.rest.authentication.service;
+
+import ddf.security.service.SecurityServiceException;
+import javax.servlet.http.HttpServletRequest;
+
+public interface AuthenticationService {
+
+  void login(HttpServletRequest request, String username, String password, String prevurl)
+      throws SecurityServiceException;
+}
