+++ ui/packages/catalog-ui-search/src/main/webapp/component/user/user.view.js
@@ -68,7 +68,7 @@ module.exports = Marionette.ItemView.extend({
 
         $.ajax({
             type: "POST",
-            url: "/services/login",
+            url: "./internal/login",
             data: {
                 "username": view.$('#username').val(),
                 "password": view.$('#password').val(),
